Create IO-based evaluator, and remove other effect types
