+++ src/main/scala/com/hhandoko/aktiform/app/controller/FormController.scala
@@ -1,15 +1,19 @@
 package com.hhandoko.aktiform.app.controller
 
 import java.util.{Map => JMap}
-import scala.concurrent.duration._
-import scala.concurrent.{Await, Future}
-import scala.util.Try
 
 import cats.effect.IO
 import io.circe.Json
 import org.graalvm.polyglot.Context
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.web.bind.annotation.{GetMapping, PathVariable, PostMapping, RequestParam, ResponseBody, RestController}
+import org.springframework.web.bind.annotation.{
+  GetMapping,
+  PathVariable,
+  PostMapping,
+  RequestParam,
+  ResponseBody,
+  RestController
+}
 
 import com.hhandoko.aktiform.api.html.input.{Form, InputNumberField, InputTextAreaField, InputTextField}
 import com.hhandoko.aktiform.api.html.{Page, Section}
@@ -17,6 +21,7 @@ import com.hhandoko.aktiform.api.task.Step
 import com.hhandoko.aktiform.app.config.ResourcesConfig
 import com.hhandoko.aktiform.app.view.render.HtmlBootstrapRender
 import com.hhandoko.aktiform.core.Capabilities
+import com.hhandoko.aktiform.core.runtime.Evaluator
 
 @RestController
 final class FormController @Autowired() (
@@ -50,41 +55,9 @@ final class FormController @Autowired() (
     val filledForm  = form(id).fill(data.asScala.toMap)
     val formPayload = filledForm.toJson
 
-    val printStep     = PrintStep.run(formPayload)
-    val transformStep = TransformStep.run(printStep)
-
-    val stepsSeq = List(PrintStep.run _, TransformStep.run _)
-    val stepSequence = stepsSeq.foldLeft(formPayload) {
-      case (acc, step) => step(acc)
-    }
-
-    val stepsTryRec = List(Try(PrintStep.run _), Try(TransformStep.run _))
-    val stepTryRecursive =
-      recursive[Json, Try](
-        formPayload,
-        stepsTryRec,
-        (acc, eff) => eff.map(fun => fun(acc)),
-        eff => eff.toEither.left.map(_ => "Fail")
-      ).fold(err => Json.fromString(err), identity)
-
-    import scala.concurrent.ExecutionContext.Implicits.global
-    val stepsFutRec = List(Future(PrintStep.run _), Future(TransformStep.run _))
-    val stepFutRecursive =
-      recursive[Json, Future](
-        formPayload,
-        stepsFutRec,
-        (acc, eff) => eff.map(fun => fun(acc)),
-        eff => Await.result(eff.map(res => Right(res)).recover { case _ => Left("Fail") }, 30.seconds)
-      ).fold(err => Json.fromString(err), identity)
-
-    val stepsIORec = List(IO(PrintStep.run _), IO(TransformStep.run _))
-    val stepIORecursive =
-      recursive[Json, IO](
-        formPayload,
-        stepsIORec,
-        (acc, eff) => eff.map(fun => fun(acc)),
-        eff => eff.attempt.unsafeRunSync().left.map(_ => "Fail")
-      ).fold(err => Json.fromString(err), identity)
+    val stepsIORec    = List(IO(PrintStep.run _), IO(TransformStep.run _))
+    val stepsIOEval   = Evaluator.run(stepsIORec) _
+    val stepsIOResult = stepsIOEval(formPayload).fold(Json.fromString, identity)
 
     data.asScala
       .map { case (key, value) => s"$key -> $value" }
@@ -92,40 +65,8 @@ final class FormController @Autowired() (
       .prependedAll("<h1>Form Data</h1>")
       .prependedAll(s"id -> $id")
       .prependedAll("<h1>Request</h1>")
-      .concat("<h1>Print Step</h1>")
-      .concat(printStep.spaces4SortKeys)
-      .concat("<h1>Transform Step</h1>")
-      .concat(transformStep.spaces4SortKeys)
-      .concat("<h1>Combined</h1>")
-      .concat(stepSequence.spaces4SortKeys)
-      .concat("<h1>Recursive (Try)</h1>")
-      .concat(stepTryRecursive.spaces4SortKeys)
-      .concat("<h1>Recursive (Future)</h1>")
-      .concat(stepFutRecursive.spaces4SortKeys)
       .concat("<h1>Recursive (IO)</h1>")
-      .concat(stepIORecursive.spaces4SortKeys)
-  }
-
-  private def recursive[A, F[_]](
-      acc: A,
-      steps: List[F[A => A]],
-      map: (A, F[A => A]) => F[A],
-      eval: F[A] => Either[String, A]
-  ): Either[String, A] = {
-    steps match {
-      case Nil =>
-        Right(acc)
-
-      case step :: Nil =>
-        eval(map(acc, step))
-
-      case step :: tail =>
-        eval(map(acc, step))
-          .fold(
-            err => Left(err),
-            res => recursive(res, tail, map, eval)
-          )
-    }
+      .concat(stepsIOResult.spaces4SortKeys)
   }
 
   object PrintStep extends Step {
